tools/requirements.txt
black >= 23.7.0
pylint >= 2.17.5
numpy >= 1.25.2
opencv-python >= 4.8.0.74
Pillow >= 10.0.0